refactor(output): move all output related features in output module
