+++ tests/study/output/test_output_service.py
