+++ tests/study/web/test_output_blueprint.py
@@ -11,7 +11,8 @@
 # This file is part of the Antares project.
 import pytest
 
-from antarest.study.business.output.variables_management import (
+from antarest.study.output.output_model import OutputVariablesType
+from antarest.study.output.variables_management import (
     AreaOutputId,
     LinkOutputId,
     OutputItemId,
@@ -19,7 +20,6 @@ from antarest.study.business.output.variables_management import (
     ShortTermStorageOutputId,
     ThermalClusterOutputId,
 )
-from antarest.study.output.output_model import OutputVariablesType
 from antarest.study.web.output_blueprint import _to_item_id
